k8s: Trigger endpoint delete from k8s delete events
In a user environment, we encountered a scenario where pods would
"disappear" and Cilium seemingly received no notification of deletion
(eg, CNI DELETE command). If this occurs regularly over time, then
this leads to Cilium managing a series of phantom endpoints which no
longer have corresponding k8s objects.

Some symptoms include:

* Exhaustion of IP address management pool, causing inability to deploy
  new endpoints on the node;
* Metrics such as cilium_endpoint_count increase to the maximum
  number of endpoints on the node (typically limited by IP pool); and
* Label resolution controllers reporting errors in "cilium status"
  output, such as: pod.core "my-pod-name" not found.

Fix this by also monitoring for k8s pod deletion events and attempting
to match those pod deletions against the local list of endpoints. If the
corresponding endpoint is found, delete it to clean up the corresponding
resources.

As a result, the endpoint may in fact be cleaned up via either a CNI
DELETE or a k8s pod delete event; in either case, whichever event comes
second will ignore the request and report success if the endpoint cannot
be found.

To find out how often this case is hit, and assist potential debugging
of such scenarios, the number of endpoints that are deleted via this
mechanism is tracked by a new metric, "cilium_endpoint_gc_count".

To ensure that this new endpoint deletion path doesn't block handling of
other kubernetes events, this functionality is run from a dedicated
goroutine. A new metric, "cilium_endpoint_active_gc_count" tracks the
current number of goroutines that are attempting to garbage collect
endpoints corresponding to deleted pods. If this metric is ever above
zero for an extended period, then it likely means that something is
stuck and further debugging should be performed (with "gops stack"
output).

There's also some potential race conditions between CNI ADD, DELETE, ADD
and k8s pod delete events for newly created pods with the same name, in
which case we want to avoid inadvertently deleting endpoints
corresponding to newly created pods. See findEndpointForDeletedPod() for
a more detailed explanation of the cases we try to handle there. This
behaviour can be tweaked by a new hidden agent argument,
"endpoint-gc-threshold". Most users should not need to tweak this
setting, but it is exposed and bound to an environment variable in case
it does need to be modified in some environment.

Signed-off-by: Joe Stringer <joe@cilium.io>