k8s: Trigger pod deletion for jobs with no running pods
This commit is derived from b3adc4d9410c ("k8s: delete IPs from ipcache
for no running Pods"), with a minimal backport to branch v1.7.

The specific pieces pulled in here are to check for k8s pod add or
update events that cause the pod to go into a Succeeded or Failed phase,
indicating that the pod is no longer running. In these cases, we want to
trigger the deletion path to ensure garbage collection of Cilium
Endpoint objects to prevent those pods from continuing to consume
resources on the node.

Signed-off-by: Joe Stringer <joe@cilium.io>